Extract the Property class into a Database container
